Put cache code in separate file, track lastSent and lastSeen separately.
